+++ src/test/java/domain/position/ColumnTest.java
@@ -1,8 +1,7 @@
-package domain;
+package domain.position;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.position.Column;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
