+++ src/test/java/domain/position/RowTest.java
@@ -1,9 +1,8 @@
-package domain;
+package domain.position;
 
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.position.Row;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
